bitcoin_crew/requirements.txt
crewai
langchain-groq
python-dotenv
requests 